+++ src/components/timetable/Grid/LectureGrid.tsx
@@ -1,88 +1,42 @@
 import { css } from '@styled-system/css'
+import { useMemo } from 'react'
 
 import { GetTimetableByTimetableIdResponse } from '@/api/types/timetable'
-import LectureSticker from '@/components/timetable/Grid/LectureSticker'
-import { TimeCell } from '@/components/timetable/Grid/TimetableLayout'
-import { COLOR_INFO } from '@/lib/constants/timetableColors'
+import NoScheduledArea from '@/components/timetable/Grid/NoScheduledArea'
+import ScheduledArea from '@/components/timetable/Grid/ScheduledArea'
 import { getWeeknTimeList, lectureDataPreprocess } from '@/util/timetableUtil'
 
 interface LectureGridProps {
-  timetableId?: number
+  timetableId?: number // 친구 시간표의 경우 undefined
   timetableData: GetTimetableByTimetableIdResponse
   isMine?: boolean
 }
 
 const LectureGrid = ({ timetableId, timetableData, isMine = false }: LectureGridProps) => {
-  const { time, week } = getWeeknTimeList(timetableData.courses, timetableData.schedules)
+  const { time, week } = useMemo(
+    () => getWeeknTimeList(timetableData.courses, timetableData.schedules),
+    [timetableData],
+  )
   const weekCnt = week.length
   const timeCnt = time.length - 1
 
-  const { lectureGrid } = lectureDataPreprocess(timetableData.courses, timetableData.schedules, weekCnt, timeCnt)
+  const { lectureGrid, noScheduled } = useMemo(
+    () => lectureDataPreprocess(timetableData.courses, timetableData.schedules, weekCnt, timeCnt),
+    [timetableData, weekCnt, timeCnt],
+  )
   const colorTheme = timetableData.color
-  let lecCnt = 0
 
   return (
-    <div
-      className={css({
-        display: 'flex',
-        flexDir: 'row',
-        borderLeft: '1px solid {colors.lightGray.1}',
-        roundedBottom: 10,
-        bgColor: 'white',
-      })}
-    >
-      <div className={css({ display: 'flex', flexDir: 'column' })}>
-        {time.map((val, index) => {
-          return (
-            <div
-              key={index}
-              className={TimeCell({
-                sidebar: true,
-                header: index === 0,
-                end: index === time.length - 1 ? 'leftEnd' : undefined,
-              })}
-            >
-              {val}
-            </div>
-          )
-        })}
-      </div>
-      <div className={css({ display: 'flex', flexDir: 'column', flex: 1 })}>
-        <div className={css({ display: 'flex', flexDir: 'row' })}>
-          {week.map((days, index) => {
-            return (
-              <div key={index} className={css({ flex: 1 }, TimeCell.raw({ header: true }))}>
-                {days}
-              </div>
-            )
-          })}
-        </div>
-        <div className={css({ display: 'grid' })} style={{ gridTemplateColumns: `repeat(${weekCnt}, 1fr)` }}>
-          {lectureGrid.map((lectures, gridInd) => {
-            return (
-              <div
-                key={gridInd}
-                className={TimeCell({
-                  lectureGrid: true,
-                  end: gridInd === weekCnt * timeCnt - 1 ? 'rightEnd' : undefined,
-                })}
-              >
-                {lectures.map((lecture, lecInd) => {
-                  return (
-                    <LectureSticker
-                      key={lecInd}
-                      timetableId={timetableId}
-                      data={lecture}
-                      bgColor={COLOR_INFO[colorTheme]['rand'][lecCnt++ % COLOR_INFO[colorTheme]['rand'].length]}
-                      isMine={isMine}
-                    />
-                  )
-                })}
-              </div>
-            )
-          })}
-        </div>
-      </div>
+    <div className={css({ display: 'flex', flexDir: 'column', gap: { base: 2.5, smDown: 1 } })}>
+      <ScheduledArea
+        data={lectureGrid}
+        colorTheme={colorTheme}
+        timetableId={timetableId}
+        isMine={isMine}
+        time={time}
+        week={week}
+      />
+      {noScheduled.length !== 0 && <NoScheduledArea data={noScheduled} timetableId={timetableId} isMine={isMine} />}
     </div>
   )
 }
